Update packet format
